After the transaction is complete to the audit log

+++ lib/pf/UnifiedApi.pm
@@ -26,6 +26,7 @@ use JSON::MaybeXS qw();
 use pf::services;
 use pf::pfqueue::status_updater::redis;
 use pf::util::pfqueue qw(consumer_redis_client);
+use Scalar::Util qw(weaken);
 
 use Mojo::Base 'Mojolicious';
 use pf::util qw(add_jitter rss_kb);
@@ -175,20 +176,25 @@ after_dispatch_cb
 
 sub after_dispatch_cb {
     my ($c) = @_;
-    eval {
-        $c->audit_request if $c->can("audit_request");
-    };
+    weaken(my $weak_c = $c);  # break the cycle
+    $c->tx->on(finish => sub {
+        return unless $weak_c;  # guard: check it wasn't garbage collected
+        my $app = $weak_c->app;
 
-    if($@) {
-        $c->log->error("Failed to audit request: $@");
-    }
+        eval {
+            $weak_c->audit_request if $weak_c->can("audit_request");
+        };
 
-    my $app = $c->app;
-    my $max = $app->{max_requests_handled} //= add_jitter( $MAX_REQUEST_HANDLED, $REQUEST_HANDLED_JITTER );
-    my $max_rss_kb = ($Config{advanced}{pfperl_api_max_rss} // 1024) * 1024;
-    if (++$app->{requests_handled} >= $max || ($max_rss_kb > 0 && rss_kb() > $max_rss_kb)) {
-        kill 'QUIT', $$;
-    }
+        if($@) {
+            $weak_c->log->error("Failed to audit request: $@");
+        }
+
+        my $max = $app->{max_requests_handled} //= add_jitter( $MAX_REQUEST_HANDLED, $REQUEST_HANDLED_JITTER );
+        my $max_rss_kb = ($Config{advanced}{pfperl_api_max_rss} // 1024) * 1024;
+        if (++$app->{requests_handled} >= $max || ($max_rss_kb > 0 && rss_kb() > $max_rss_kb)) {
+            kill 'QUIT', $$;
+        }
+    });
 
     $c->after_dispatch;
     return;
